Async POSIX signal handler
Commit a067ab2: feat: move to async POSIX signal handler
Instead of trying to handle signals for every step of the `cursive`
event loop, move the signal handling into its own asynchronous task and
send callbacks to `cursive` when a signal arrives.

PR 1305: fix: restore SIGHUP in signal handler
Restore the SIGHUP signal handler which was accidentally removed in a067ab2.

The commit had to be reverted and the PR was closed.

Commit a1a9863: Revert "feat: move to async POSIX signal handler"

ThomasFrans commented: From what I can see, it's an issue introduced by the
crossterm backend. The ncurses one doesn't keep the process running. Currently
I suspect Cursive is doing a write to /dev/tty which is closed, and that write
blocks on the crossterm backend and returns immediately on the ncurses backend.
I guess it could be resolved by switching the default backend (again)?
This would solve #1309 and #1310 and #1282. I'd love to hear thoughts
on this idea. If it's an OK solution, I'd try to see whether I can open an
issue at Cursive or Crossterm to resolve the issue there, and meanwhile switch
backend and add the SIGHUP handler back to finally close all these issues

A commit to deprecate the crossterm/pancurses/termion backends will follow.