Refactored all config/cache/runtime directory lookups
Kill the CLI options for base directory name and configuration file name.

Standardized the way that user directories are lookup up and used.
This includes the replacement of the 'platform_dirs' crate with the 'dirs'
crate.

Because there is less customizability for files, crate interdependencies
and global states could be removed. This greatly reduces the number of
event loops that can return an error. Many code branches are killed as
unnecessary.